chromium/third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/cache-storage-reporting-dedicated-worker.https-expected.txt

This is a testharness.js-based test.
FAIL COEP support on DedicatedWorker. assert_equals: expected "https://web-platform.test:8444/html/cross-origin-embedder-policy/resources/universal-worker.js" but got "https://web-platform.test:8444/html/cross-origin-embedder-policy/resources/iframe.html"
FAIL COEP-Report-Only support on DedicatedWorker. assert_equals: expected "https://web-platform.test:8444/html/cross-origin-embedder-policy/resources/universal-worker.js" but got "https://web-platform.test:8444/html/cross-origin-embedder-policy/resources/iframe.html"
Harness: the test ran to completion.
